Migrate oxide_override to its own module
We're about to expand the logic around `OxideOverride`, move it into its
own module now for a more understandable diff.
